+++ scripts/detect_bad_deps.js
@@ -12,7 +12,7 @@
 import madge from "madge";
 import { exec } from "child_process";
 
-const MAX_CYCLE_LEN = 57; // NEVER EVER increase this value
+const MAX_CYCLE_LEN = 56; // NEVER EVER increase this value
 
 const cmd = "flow check --profile --merge-timeout 0";
 exec(cmd, function(error, stdout, stderr) {
