+++ src/serializer/ResidualFunctionInitializers.js
@@ -12,7 +12,7 @@
 import { FunctionValue, Value } from "../values/index.js";
 import * as t from "@babel/types";
 import type { BabelNodeStatement } from "@babel/types";
-import { NameGenerator } from "../utils/PreludeGenerator.js";
+import { NameGenerator } from "../utils/NameGenerator.js";
 import traverseFast from "../utils/traverse-fast.js";
 import invariant from "../invariant.js";
 import { voidExpression, nullExpression } from "../utils/babelhelpers.js";
