+++ src/utils/PreludeGenerator.js
@@ -17,7 +17,7 @@ import type {
   BabelNodeStatement,
   BabelNodeMemberExpression,
 } from "@babel/types";
-import invariant from "../invariant.js";
+import { NameGenerator } from "./NameGenerator.js";
 
 export class PreludeGenerator {
   constructor(debugNames: ?boolean, uniqueSuffix: ?string) {
@@ -95,53 +95,5 @@ export class PreludeGenerator {
     this.prelude.push(t.variableDeclaration("var", [t.variableDeclarator(ref, init)]));
     this.memoizedRefs.set(key, ref);
     return ref;
-  }
-}
-
-function escapeInvalidIdentifierCharacters(s: string): string {
-  let res = "";
-  for (let c of s)
-    if ((c >= "0" && c <= "9") || (c >= "a" && c <= "z") || (c >= "A" && c <= "Z")) res += c;
-    else res += "_" + c.charCodeAt(0);
-  return res;
-}
-
-const base62characters = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-function base62encode(n: number): string {
-  invariant((n | 0) === n && n >= 0);
-  if (n === 0) return "0";
-  let s = "";
-  while (n > 0) {
-    let f = n % base62characters.length;
-    s = base62characters[f] + s;
-    n = (n - f) / base62characters.length;
-  }
-  return s;
-}
-
-export class NameGenerator {
-  constructor(forbiddenNames: Set<string>, debugNames: boolean, uniqueSuffix: string, prefix: string) {
-    this.prefix = prefix;
-    this.uidCounter = 0;
-    this.debugNames = debugNames;
-    this.forbiddenNames = forbiddenNames;
-    this.uniqueSuffix = uniqueSuffix;
-  }
-  prefix: string;
-  uidCounter: number;
-  debugNames: boolean;
-  forbiddenNames: Set<string>;
-  uniqueSuffix: string;
-  generate(debugSuffix: ?string): string {
-    let id;
-    do {
-      id = this.prefix + base62encode(this.uidCounter++);
-      if (this.uniqueSuffix.length > 0) id += this.uniqueSuffix;
-      if (this.debugNames) {
-        if (debugSuffix) id += "_" + escapeInvalidIdentifierCharacters(debugSuffix);
-        else id += "_";
-      }
-    } while (this.forbiddenNames.has(id));
-    return id;
   }
 }
